+++ crates/gen-bsky/src/draft/blog_post/front_matter/bluesky.rs
@@ -1,31 +1,31 @@
 use serde::Deserialize;
 
 #[derive(Default, Debug, Clone, Deserialize)]
-pub struct Taxonomies {
-    #[allow(dead_code)]
-    tags: Vec<String>,
+pub(crate) struct Bluesky {
+    description: Option<String>,
+    tags: Option<Vec<String>>,
 }
 
-impl Taxonomies {
-    pub fn new(tags: Vec<String>) -> Self {
-        Taxonomies { tags }
+impl Bluesky {
+    pub(crate) fn description(&self) -> &str {
+        self.description.as_deref().unwrap_or("")
     }
 
-    pub fn tags(&self) -> &Vec<String> {
-        self.tags.as_ref()
+    pub(crate) fn tags(&self) -> Vec<String> {
+        self.tags.clone().unwrap_or_default()
     }
 
-    pub fn hashtags(&self) -> Vec<String> {
+    pub(crate) fn hashtags(&self) -> Vec<String> {
         let mut hashtags = vec![];
-        for tag in self.tags() {
+        for tag in &self.tags() {
             // convert tag to hashtag by capitalising the first letter of each word, removing the spaces and prefixing with a # if required
             let formatted_tag = tag
                 .split_whitespace()
                 .map(|word| {
                     let mut chars = word.chars();
                     match chars.next() {
-                        Some(first) => first.to_uppercase().collect::<String>() + chars.as_str(),
                         None => String::new(),
+                        Some(first) => first.to_uppercase().collect::<String>() + chars.as_str(),
                     }
                 })
                 .collect::<String>();