+++ crates/pcu/src/cli/bsky/commands/cmd_draft.rs
@@ -1,5 +1,7 @@
 mod site_config;
 
+use std::path::PathBuf;
+
 use clap::Parser;
 use config::Config;
 use gen_bsky::Draft;
@@ -15,7 +17,7 @@ pub struct CmdDraft {
     #[arg(short, long)]
     pub filter: Option<String>,
     /// Optional path to file or directory of blog post(s) to process
-    pub path: Option<String>,
+    pub paths: Vec<PathBuf>,
     /// Optional date from which to process blog post(s)
     /// Date format: YYYY-MM-DD
     /// Default: Current date
@@ -27,37 +29,36 @@ pub struct CmdDraft {
 }
 
 impl CmdDraft {
-    pub async fn run(&self, client: &Client, settings: &Config) -> Result<CIExit, Error> {
+    pub async fn run(&mut self, client: &Client, settings: &Config) -> Result<CIExit, Error> {
         // find the potential file in the git repo
 
         let base_url = SiteConfig::new()?.base_url();
         let store = &settings.get_string("store")?;
-        let path = if let Some(path) = self.path.as_ref() {
-            path
-        } else {
-            DEFAULT_PATH
+        if !self.paths.is_empty() {
+            self.paths.push(PathBuf::from(DEFAULT_PATH))
         };
 
         log::trace!(
-            "Key parameters:\n\tBase:\t`{base_url}`\n\tstore:\t`{store}`\n\tpath:\t`{path}`"
+            "Key parameters:\n\tBase:\t`{base_url}`\n\tstore:\t`{store}`\n\tpath:\t`{:#?}`",
+            self.paths
         );
 
-        let mut posts_builder = Draft::builder();
-        let posts_res = posts_builder
-            .with_base_url(&base_url)
-            .with_store(store)
-            .add_path_or_file(path)?
+        let mut builder = Draft::builder(base_url);
+
+        // Add the paths specified at the command line.
+        for path in self.paths.iter() {
+            builder.add_path_or_file(path)?;
+        }
+
+        // Set the filters to qualify the blog posts
+        builder
             .with_minimum_date(self.date)?
-            .with_allow_draft(self.allow_draft)
-            .build();
+            .with_allow_draft(self.allow_draft);
 
-        let Ok(mut posts) = posts_res else {
-            log::warn!("{}", posts_res.err().unwrap());
-            return Ok(CIExit::DraftedForBluesky);
-        };
+        let mut posts = builder.build().await?;
 
-        posts.process_posts().await?;
-        posts.write_bluesky_posts()?;
+        posts.write_referrers(None)?;
+        posts.write_bluesky_posts(None)?;
 
         let sign = Sign::Gpg;
         // Commit the posts to the git repo