+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 👷 ci(circleci)-enhance check_last_commit workflow(pr [#646])
 - 👷 ci(config)-update CircleCI workflow filters(pr [#647])
 - 📝 docs(README)-update project title casing(pr [#649])
+- Simplify-and-improve-draft-structure(pr [#650])
 
 ## [0.4.56] - 2025-07-30
 
@@ -1625,6 +1626,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#647]: https://github.com/jerus-org/pcu/pull/647
 [#648]: https://github.com/jerus-org/pcu/pull/648
 [#649]: https://github.com/jerus-org/pcu/pull/649
+[#650]: https://github.com/jerus-org/pcu/pull/650
 [Unreleased]: https://github.com/jerus-org/pcu/compare/v0.4.56...HEAD
 [0.4.56]: https://github.com/jerus-org/pcu/compare/v0.4.55...v0.4.56
 [0.4.55]: https://github.com/jerus-org/pcu/compare/v0.4.54...v0.4.55